code/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu116
salesforce-lavis==1.0.2
torch==1.13.1+cu116
torchaudio==0.13.1+cu116
torchvision==0.14.1+cu116
transformers==4.26.1
Pillow==9.4.0
numpy==1.24.2
fastapi==0.93.0
uvicorn==0.20.0
pydantic==1.10.6
requests==2.22.0
wget==3.2
python-multipart==0.0.6
